# Speed up CDN purges by re-using a pool manager