Helps https://github.com/python/docsbuild-scripts/issues/169.

Looking at the server logs (see [en-3.12.log](https://github.com/user-attachments/files/16576053/en-3.12.log)) for building the `en/3.12` documentation, I noticed that during the publishing stage, there are 1,612 instances of purging URLs from the CDN. It does both the `/3/` and `/3.12/` URLs.

For example, it begins:

```
2024-08-11 08:33:44,746 DEBUG en/3.12: Purging https://docs.python.org/3/ from CDN
2024-08-11 08:33:44,747 DEBUG en/3.12: Starting new HTTPS connection (1): docs.python.org:443
2024-08-11 08:33:44,796 DEBUG en/3.12: https://docs.python.org:443 "PURGE /3/ HTTP/1.1" 200 53
2024-08-11 08:33:44,798 DEBUG en/3.12: Purging https://docs.python.org/3.12/ from CDN
2024-08-11 08:33:44,800 DEBUG en/3.12: Starting new HTTPS connection (1): docs.python.org:443
2024-08-11 08:33:44,841 DEBUG en/3.12: https://docs.python.org:443 "PURGE /3.12/ HTTP/1.1" 200 53
2024-08-11 08:33:44,843 DEBUG en/3.12: Purging https://docs.python.org/3/.buildinfo from CDN
2024-08-11 08:33:44,844 DEBUG en/3.12: Starting new HTTPS connection (1): docs.python.org:443
2024-08-11 08:33:44,882 DEBUG en/3.12: https://docs.python.org:443 "PURGE /3/.buildinfo HTTP/1.1" 200 52
2024-08-11 08:33:44,885 DEBUG en/3.12: Purging https://docs.python.org/3/about.html from CDN
2024-08-11 08:33:44,886 DEBUG en/3.12: Starting new HTTPS connection (1): docs.python.org:443
2024-08-11 08:33:44,926 DEBUG en/3.12: https://docs.python.org:443 "PURGE /3/about.html HTTP/1.1" 200 53
```

And ends:
```
2024-08-11 08:35:00,565 DEBUG en/3.12: Purging https://docs.python.org/3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2 from CDN
2024-08-11 08:35:00,566 DEBUG en/3.12: Starting new HTTPS connection (1): docs.python.org:443
2024-08-11 08:35:00,606 DEBUG en/3.12: https://docs.python.org:443 "PURGE /3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2 HTTP/1.1" 200 53
```

This takes 2 minutes and 44 seconds.

Each `purge()` command is calling `requests.request()`, which involves starting new HTTPS connections for each and every URL. The "Starting new HTTPS connection" [comes from urllib3](https://github.com/urllib3/urllib3/blob/d494c47cbd6a82a94de02621426c973b562ceb86/src/urllib3/connectionpool.py#L1052), the underlying HTTP client used by Requests.

We can do better by creating a Requests "session" which will open a single connection, and re-use it by calling `session.request()` in `purge()`:

https://requests.readthedocs.io/en/latest/user/advanced/

Or, because Requests is essentially wrapping the urllib3 HTTP client, we might as well use that urllib3 directly instead of Requests. The API is virtually the same.

# Comparison

Re-using a connection makes it about three times quicker on my machine.

Timings from my Mac using the following test scripts.

## Original `purge()`: 33 seconds

```python
import logging
from build_docs import purge

PATHS = [  # the 1,612 paths
    "3/",
    <snip>
    "3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2",
]

logging.basicConfig(level=logging.DEBUG)

for path in PATHS:
    purge(path)
```

## With `requests.Session()`: 11 seconds

```python
import logging
import requests  # NEW
from build_docs import purge

PATHS = [  # the 1,612 paths
    "3/",
    <snip>
    "3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2",
]

logging.basicConfig(level=logging.DEBUG)

session = requests.Session()  # NEW

for path in PATHS:
    purge(session, path)  # CHANGED
```
Output is like:
```
DEBUG:root:Purging https://docs.python.org/3.12/archives/python-3.12.0b4-docs-html.tar.bz2 from CDN
DEBUG:urllib3.connectionpool:https://docs.python.org:443 "PURGE /3.12/archives/python-3.12.0b4-docs-html.tar.bz2 HTTP/1.1" 200 54
DEBUG:root:Purging https://docs.python.org/3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2 from CDN
DEBUG:urllib3.connectionpool:https://docs.python.org:443 "PURGE /3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2 HTTP/1.1" 200 54
```

## With `urllib3.PoolManager()`: 10 seconds

```python
import logging
import urllib3  # CHANGED
from build_docs import purge

PATHS = [  # the 1,612 paths
    "3/",
    <snip>
    "3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2",
]

logging.basicConfig(level=logging.DEBUG)

http = urllib3.PoolManager()  # CHANGED

for path in PATHS:
    purge(http, path)  # CHANGED
```
Output is like:
```
DEBUG:root:Purging https://docs.python.org/3.12/archives/python-3.12.0b4-docs-html.tar.bz2 from CDN
DEBUG:urllib3.connectionpool:https://docs.python.org:443 "PURGE /3.12/archives/python-3.12.0b4-docs-html.tar.bz2 HTTP/1.1" 200 0
DEBUG:root:Purging https://docs.python.org/3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2 from CDN
DEBUG:urllib3.connectionpool:https://docs.python.org:443 "PURGE /3.12/archives/python-3.12.3-docs-pdf-a4.tar.bz2 HTTP/1.1" 200 0
```