@hugovk 2024-09-18
The improvement is even larger on the docs server. For `en/3.14`: 63s -> 3.5s 🚀 

# Before: 63s

```
64871:2024-09-17 12:32:39,535 DEBUG en/3.14: Purging https://docs.python.org/dev/ from CDN
64872:2024-09-17 12:32:39,538 DEBUG en/3.14: Starting new HTTPS connection (1): docs.python.org:443
64873:2024-09-17 12:32:39,580 DEBUG en/3.14: https://docs.python.org:443 "PURGE /dev/ HTTP/1.1" 200 52
64874:2024-09-17 12:32:39,582 DEBUG en/3.14: Purging https://docs.python.org/3.14/ from CDN
64875:2024-09-17 12:32:39,583 DEBUG en/3.14: Starting new HTTPS connection (1): docs.python.org:443
64876:2024-09-17 12:32:39,624 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/ HTTP/1.1" 200 53
64877:2024-09-17 12:32:39,626 DEBUG en/3.14: Purging https://docs.python.org/dev/.buildinfo from CDN
64878:2024-09-17 12:32:39,627 DEBUG en/3.14: Starting new HTTPS connection (1): docs.python.org:443
64879:2024-09-17 12:32:39,668 DEBUG en/3.14: https://docs.python.org:443 "PURGE /dev/.buildinfo HTTP/1.1" 200 52
64880:2024-09-17 12:32:39,670 DEBUG en/3.14: Purging https://docs.python.org/dev/404.html from CDN

[snip 3328 lines]

68209:2024-09-17 12:33:36,345 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs-pdf-a4.tar.bz2 HTTP/1.1" 200 52
68210:2024-09-17 12:33:36,347 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs-html.tar.bz2 from CDN
68211:2024-09-17 12:33:36,348 DEBUG en/3.14: Starting new HTTPS connection (1): docs.python.org:443
68212:2024-09-17 12:33:36,389 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs-html.tar.bz2 HTTP/1.1" 200 52
68213:2024-09-17 12:33:36,391 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs-text.tar.bz2 from CDN
68214:2024-09-17 12:33:36,392 DEBUG en/3.14: Starting new HTTPS connection (1): docs.python.org:443
68215:2024-09-17 12:33:36,432 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs-text.tar.bz2 HTTP/1.1" 200 50
68216:2024-09-17 12:33:36,434 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs.epub from CDN
68217:2024-09-17 12:33:36,436 DEBUG en/3.14: Starting new HTTPS connection (1): docs.python.org:443
68218:2024-09-17 12:33:36,493 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs.epub HTTP/1.1" 200 52
```

# After: 3.5s

```
36729:2024-09-18 10:39:56,295 DEBUG en/3.14: Purging https://docs.python.org/dev/ from CDN
36730:2024-09-18 10:39:56,296 DEBUG en/3.14: Resetting dropped connection: docs.python.org
36731:2024-09-18 10:39:56,343 DEBUG en/3.14: https://docs.python.org:443 "PURGE /dev/ HTTP/11" 200 0
36732:2024-09-18 10:39:56,344 DEBUG en/3.14: Purging https://docs.python.org/3.14/ from CDN
36733:2024-09-18 10:39:56,347 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/ HTTP/11" 200 0
36734:2024-09-18 10:39:56,347 DEBUG en/3.14: Purging https://docs.python.org/dev/.buildinfo from CDN
36735:2024-09-18 10:39:56,350 DEBUG en/3.14: https://docs.python.org:443 "PURGE /dev/.buildinfo HTTP/11" 200 0
36736:2024-09-18 10:39:56,350 DEBUG en/3.14: Purging https://docs.python.org/dev/404.html from CDN
36737:2024-09-18 10:39:56,353 DEBUG en/3.14: https://docs.python.org:443 "PURGE /dev/404.html HTTP/11" 200 0
36738:2024-09-18 10:39:56,353 DEBUG en/3.14: Purging https://docs.python.org/dev/about.html from CDN

[snip 2221]

38960:2024-09-18 10:39:59,833 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs-texinfo.zip from CDN
38961:2024-09-18 10:39:59,835 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs-texinfo.zip HTTP/11" 200 0
38962:2024-09-18 10:39:59,835 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs-pdf-a4.tar.bz2 from CDN
38963:2024-09-18 10:39:59,837 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs-pdf-a4.tar.bz2 HTTP/11" 200 0
38964:2024-09-18 10:39:59,837 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs-html.tar.bz2 from CDN
38965:2024-09-18 10:39:59,840 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs-html.tar.bz2 HTTP/11" 200 0
38966:2024-09-18 10:39:59,840 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs-text.tar.bz2 from CDN
38967:2024-09-18 10:39:59,842 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs-text.tar.bz2 HTTP/11" 200 0
38968:2024-09-18 10:39:59,842 DEBUG en/3.14: Purging https://docs.python.org/3.14/archives/python-3.14.0a0-docs.epub from CDN
38969:2024-09-18 10:39:59,846 DEBUG en/3.14: https://docs.python.org:443 "PURGE /3.14/archives/python-3.14.0a0-docs.epub HTTP/11" 200 0

```